[core] Even more aggressively deduplicate the results of visibility transforms. (#7295)
fix https://github.com/microsoft/typespec/issues/7294
This changes the visibility mutators to use replacement functions
instead of mutations, making it so that if the visibility transform does
not actually modify a model, it will return the same input model
instance. Integrated with the caching system, this should produce a
minimally-copied tree. The only node that is guaranteed to be duplicated
is the root node, since we cannot deduplicate the template instance.