Remove `literalContext` (#6744)
Fixes https://github.com/Azure/autorest.csharp/issues/4746

We have a very complicated logic to give names to literal types (or
constant types from TCGC) so that we could convert them into extensible
enums in our generated code.

But actually, TCGC already has the ability to make up a name for the
constant types, which in most cases is the same as the one we have.

Therefore, to simplify our emitter, we should remove this
`literalContext` and related logic, just use the name we have from TCGC.

During making this change, some requirements come up for TCGC as well, I
will open issues, and fix them in TCGC in a near future.

Issues created for TCGC:
- https://github.com/Azure/typespec-azure/issues/2572
- https://github.com/Azure/typespec-azure/issues/2563