Leverage LSP to do compile instead of CLI (#7239)
Fix https://github.com/microsoft/typespec/issues/6663

Now we are having many problems when triggering the compile from CLI and
handling its output in our Output window like multiple loglevel, how to
handle ```stderr``, coloring and so on.

So to resolve all these problems, let's trigger compile through LSP
instead of CLI so that we can retrieve the result of compile directly
which is more reliable and also easier to do post-processing.

---------

Co-authored-by: Timothee Guerin <timothee.guerin@outlook.com>