[http-client-csharp]: Generate model factory method from LastContractView if the only thing that changed was the parameter ordering (#7462)
This PR extends the back-compat support for model factory methods by
skipping generation of model factory methods that were present in the
previous version of the generated code if only the parameter ordering
changed in the current version. In this case, the factory method from
the previous version is used.

fixes: https://github.com/microsoft/typespec/issues/7406