Add feature flag to Java generator to disable Eclipse Language Server usage (#7444)
This PR is adding a feature flag `use-eclipse-language-server` that will
control whether the Eclipse language server is used when running cod
customizations. This flag will be temporary and the default value will
be `true`. The long-term goal is for all projects using code
customizations to set this value to false, and once that is done the
flag will be removed as well as any Eclipse language server related
code.

The rationale behind this change is that while Eclipse language server
is extremely powerful it allows for too many customizations to be used.
Code customizations were originally added for niche issues found in the
generators and was meant to have limited usage, which is what Eclipse
language server circumvents. Additionally, it requires a large download
to run and is much slower than using in-memory JavaParser capabilities.