[hsjs] Correct path parameter location and lazy helper module emit. (#7494)
Closes #7342

This adds a somewhat hacky check to see if a parameter is a path
parameter as an early escape from the logic that binds properties of the
body in requests. A more complete rework is still pending, but this
should solve the linked issue for now.

It also fixes the laziness issue with helper module emit. Now helper
modules are only emitted if they are actually used by the generated
code. This prevents an issue where a helper module could be emitted
without triggering any of the "hooks" inside the emitted code that would
cause its dependencies to be collected during scaffolding.