add the api to add the compileinclude clauses (#7526)
Contributes to https://github.com/microsoft/typespec/issues/4065

This PR is not meant to solve everything in the above issue, but this
could be a start.

I think it does align with our other parts' design that the
`CSProjWriter` contains the "data" such as the description, project
reference information, etc. Instead the "provider" should hold them and
the writer just spits them out to files.

I think instead we should move all those properties on `CSProjWriter` to
the `NewProjectScaffolding` and `CSProjWriter` could consume the
`NewProjectScaffolding` as a parameter.