Reference current type name in custom code view (#7571)
If a visitor renames a type and a customer tries to rename the same type
with custom code, the customization won't work if they are referencing
the renamed type name. They would have to know what the original type
name was (which they most likely wouldn't know). This PR fixes that
behavior so the generated type is what they can reference via custom
code.