apply underline text-decoration to inlined links and improve contrast of links in codeboxes (#7567)
Addresses a18y issues reported here:

https://webcompliance.trafficmanager.net/ScanResultsApi/GetAccessibilityScanConsolidatedReport?siteId=2fae20ee-d671-4a89-c500-08dc261f6c8b

The trend I've been noticing after doing a couple of these fixes is that
if we want to deviate from the starlight/astro styles, we really need to
develop our own theme and use it wholesale. If we only use bits and
pieces (like we were doing for link coloring), we will inevitably run
into these contrast issues again the next time starlight changes any of
their css color styles. That's largely because we are so close to the
contrast limits that even an almost imperceptible change triggers the
accessibility warnings.

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>