[openapi3] add support for parameter examples via @opExample (#7403)
Fixes https://github.com/microsoft/typespec/issues/3873

This PR adds support for generating parameter examples in openapi3. 

Notes:
From the OpenAPI spec
(https://spec.openapis.org/oas/v3.0.4.html#fixed-fields-for-use-with-schema):
> When example or examples are provided in conjunction with the schema
field, the example SHOULD match the specified schema and follow the
prescribed serialization strategy for the parameter.

The parameter-level examples are meant to be serialized as they would
appear over the wire. This means in some cases, they conflict with the
parameter schema. For example, when a header parameter is an array.

The OpenAPI spec provides a table that shows what an example should look
like for any given parameter style:
https://spec.openapis.org/oas/v3.0.4.html#style-examples


Link to inconclusive discussion on the correct way to represent example
in OpenAPI 3
https://github.com/OAI/OpenAPI-Specification/discussions/4622

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>