Fix: Config not interpolating `output-dir` in parameters (#7740)
fix Azure/typespec-azure#2895

We were resolving items one after the other and output-dir was the last
of the common values. Additionally this was also inconsitent on what was
able to be interpolated(didn't really pass the same things everytime)

This PR in the end does simplify the logic quite a bit.