expose sdkContext to avoid duplicated creation (#7773)
Contributes to https://github.com/Azure/azure-sdk-for-net/pull/50936

We will call `getArmResources` from `typespec-azure-resource-manager`
package for resource detection, but its result is compiler types. We
need a way to map the compiler types into TCGC types, which we have in
mgmt emitters.

And to avoid duplicated `createSdkContext` call from TCGC, which is a
heavy task, we need to expose the `sdkContext` created in MTG emitter.