Revert dependency upgrades (#7799)
There is some incompatibility when upgrading to 9.x BCL deps and using
the Azure generator. Since most of the azure-sdk-for-net repo is still
depending on 8.x versions, we don't need to upgrade now. I'm also
reverting the SCM upgrade just to ensure that the state is consistent to
what it was prior to https://github.com/microsoft/typespec/pull/7794 and
https://github.com/microsoft/typespec/pull/7789