[http-client-csharp] emit tsp-client artifacts directly into cloned repo (#7850)
This PR adjusts the PR creation pipeline to run the `tsp-client` command
directly into the cloned copy of the azure-sdk-for-net repo rather than
doing it outside of the cloned repo and copying the emitted files. This
is to ensure that any manually added dependencies that exist in the
emitter artifact of the cloned repo are retained.

contributes to https://github.com/Azure/azure-sdk-for-net/issues/50988