http-client-java, bug fix, param in protocol method (#7888)
bug reproduced from test on override

https://github.com/Azure/typespec-azure/blob/main/packages/azure-http-specs/specs/azure/client-generator-core/override/client.tsp#L22-L29

It generates protocol method without required parameter `param1` and
`param2`.

For DPG, the signature of protocol method should not change, regardless
on there is / isn't a convenience method (and whether there is grouping
on the convenience method).

The signature of the convenience method would be on `op.convenienceApi`.