hsjs: Fixed two bugs in JSON serialization (#7940)
This PR arises from some dogfooding of the JS server generator. It fixes
two issues with the serialization code:

- Optional properties were not being handled correctly when custom
active serializers were triggered. The fix correctly wraps serialization
expressions in undefined checks if the property is declared optional.
- `requireDeclaration` (which forces anonymous types to receive a
declaration) was sometimes being provided without an altName. In cases
where an anonymous response object was in a union, the type would be
treated _as if_ a declaration was required, but failing to provide an
`altName` caused the emitter to render the anonymous model as if it were
a type name. The change creates a "Namer" for each operation and passes
it around, which can be used to create unique names with a monotonically
incrementing counter for different altname strategies, and it uses this
to create unambiguous names for anonymous types when declarations are
required for serialization.