instruction, add link to task for http-client-java (#7928)
ref https://github.com/microsoft/typespec/issues/7925

Currently we put the prompt under
[`packages/http-client-java/.github`](https://github.com/microsoft/typespec/blob/main/packages/http-client-java/.github/copilot-instructions.md
), as `packages/http-client-java` is the workspace that Java dev would
open.

(there is similarly
[JS](https://github.com/microsoft/typespec/blob/main/packages/http-client-js/.github/copilot-instructions.md))

However, when SWE works on tasks on this repository, it seems hard for
it to find this instruction to http-client-java.
I think having a link of it from `.github/copilot-instructions.md` may
help improve the SWE agent (which the agent recognizes).
(at present, we provide it as link in issue for SWE agent to follow,
e.g. https://github.com/microsoft/typespec/issues/7926)