Fix circular reference in http (#8038)
Fix issue where loading http library in tester would get stuck in vitest

Reduce circular links between files which is causing the issue.


Main change is to not use `getHttpOperation` but
resolveOperationParameters` in the visibility provider(as the only goal
was to get the verb). This not only resolve the issue but also simplify
things as it won't be calling that visibility resolution twice(as
getHttpOperation would call it back again)