hsjs: Fixed an issue with array/record repsonse serialization (#8084)
Closes https://github.com/microsoft/typespec/issues/6818

This PR fixes an issue with HSJS where if a response body is
_implicitly_ a record or array, it would fail to serialize.

The root cause of this bug was that when serializing implicit response
models with no `@body` or `@bodyRoot`, the serializer would check to see
if the response model was "emptied" after removing all metadata. This
called `properties.every(isMetadataProperty)`, which would cause arrays
and records, having no additional properties, to fall into the "empty
response" condition.

- Reworked response serialization to explicitly check for array and
record models, and serialize them appropriately by calling
`transposeExpressionToJson` as necessary.
- Enhanced the model type differentiator to allow explicitly checking
for array-ness. This is a small enhancement that simply allows array
variants to avoid occupying the valuable "else" condition of the
differentiator -- essentially, we only get one "else" variant as a
fallback, and if it's already occupied by something the differentiator
fails, so it's important to make sure that anything we can avoid putting
in the "else" variant is explicitly tested for to increase the
compatibility of the emitter to support more unions of things.
- Slightly improved body serialization for arrays and records by calling
`transposeExpressionFromJson` instead of rendering the transposition
inline, making it slightly more robust and identical to what's emitted
for nested arrays.
- Improved namespacing of global `JSON` calls by explicitly calling
`globalThis.JSON.stringify` and `globalThis.JSON.parse` instead of just
the global JSON, preventing any entity named "JSON" from shadowing the
global.
- Implemented the Spector tests for Array and Record types, which are
all passing with these changes.

---------

Co-authored-by: Christopher Radek <14189820+chrisradek@users.noreply.github.com>
Co-authored-by: Mark Cowlishaw <1054056+markcowl@users.noreply.github.com>