Revert change to call serialization hooks for non framework types (#8146)
The purpose of the change being reverted was to allow non-framework
types to participate in the serialization hooks. Specifically, this was
needed for the case of customizing a property to something like
ResourceIdentifier. But a better approach is to add a new hook to allow
loading System.Type from generators. That way it would still be
correctly treated as a framework type. I will add that new hook in a
separate PR.