Refactor how to collect buildable types for MRWContext (#8313)
Contributes to https://github.com/Azure/azure-sdk-for-net/issues/52293

Found an issue while generating MRWContext with MPG generator, we have
input models mapping to the same known type.

https://github.com/Azure/azure-sdk-for-net/blob/398dfd72dbcc2c09e41d1782e0ae4ccbc60c4ee6/eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/Primitives/KnownManagementTypes.cs#L33-L34
We are missing properties processing from `ProxyResource` input model,
because we have two `InheritableSystemObjectModelProvider` with the same
CSharpType `Azure.ResourceManager.Models.ResourceData`. The one for
`ProxyResource` has no properties, and it has a `BaseTypeProvider`
containing the properties, since they share the same CSharpType, the
`BaseTypeProvider` got skipped.

Refactor the implementation to separate `buildableTypes` and
`buildableProviders`, so that there is no strong connection between
`CSharpType` and `TypeProvider`.