http-client-java, option client-side-validation be false by default (#8502)
We now had a real issue report from service, that property is "required"
in create, but "optional" in update.

The removal of `validate()` method in model is a breaking changes.
Considered to supply an empty `validate()` to prevent the break, but
this also affect new libs.

Hence, I'd take the chance to "break it once for all".