openapi3: correctly cache and track synthetic discriminated union envelope shapes (#8711)
Closes #8677

- Adds a `manuallyTrack` method to the usage tracker, allowing the
emitter to bind the usage of synthetic discriminated union envelopes to
the union type's usage (this influences how the emitter decides to
rename models, which it only does if a type has multiple usage).
- Caches the types of created envelopes by visibility if the union
variant is transformed by the visibility.
- If a variant is subject to visibility, but not transformed, the
default envelope model is used so that all untransformed variants will
share a schema.
- This system is just to ensure that the correct _number_ of envelope
models are created, one for each active visibility transform that will
result in a different schema.
- Adds a test against regressions that validates the envelopes are
created with the correct structure, binding to the correct transformed
models of the underlying variant types.