Add mutator framework and canonicalization engine (#8822)
This PR adds packages for a new mutation framework that is significantly
more flexible and capable than the current `mutateSubgraph` suite of
APIs, and provides an HTTP Canonicalization library that makes
implementing HTTP clients and servers significantly easier than current
state of the art. Consult the READMEs for more information!