compiler: Exclude false scope members from encoded name duplication tracking (#8681)
This is a very strange bug that likely stems from the way mutators work.

When mutating a model for visibility transforms, we create clones of the
model's properties. These clones end up checked/finished, even if they
are not actually _attached_ to the model, and vice versa (the original
properties may be checked/finished even if the clone ends up attached).

This runs stateful decorators, like `@encodedName`. Unfortunately, in
these cases, it means that both the clone and the original property end
up binding an encoded name within the same scope, which the validator
later sees as a conflict.

However, only one of these properties is actually a member of the scope
at validation time. This PR uses that to exclude "false members". These
are properties that exist in the state map (the decorator has been
applied to them), but aren't actually bound to the parent they claim to
be a member of. We can work around this mutator bug in the validator by
checking if the target is actually within the set of member values (in
this case, the properties actually present in the parent model, but this
will also work for enum members and union variants).

Closes #8599