Remove some dependencies from Java TypeSpec generator (#8939)
Makes a few small changes to what Java's TypeSpec generator uses to
write itself so that `azure-json`, `azure-xml`, and Jackson dependencies
can be excluded from the shipping package.

This was done as `azure-json` and `azure-xml` functionality is
equivalent to the built-in ClientCore JSON and XML serialization
functionality. Including both just resulted in further executable size
bloat.

Jackson shouldn't be used anymore directly due to the above
dependencies, so anything that resulted in that being used was updated
as well to use ClientCore serialization concepts.

Finally, POM exclusions were added to stop adding the packages mentioned
on the first line in the executable JAR that is shipped as part of the
Java TypeSpec generator. This reduces the shipped package size by
roughly 1-2 MB (or 10-20%) in preparation for changes that will
reintroduce using Eclipse's Java code formatting dependencies which will
reintroduce many more MB of executable size.