Use Eclipse JDT Core in-memory code formatting again (#9022)
Switch Java back to using Eclipse JDT Core to perform code formatting
rather than using Maven to run the Spotless plugin. This change is being
done as running Spotless in parallel may result in deadlocks / failures
due to multiple threads attempting to get an exclusive lock on a file
managed by Spotless. This results in intermittent failures when
validating code generation in azure-sdk-for-java. Additionally, using
Maven to run Spotless is much slower than using in-memory formatting
which should be equivalent as Spotless uses Eclipse JDT Core to format
code (except removal and sorting of imports uses Google Java Format and
basic alphabetical sorting).

This PR includes some additional changes as follows:

- Import ordering was being done by logic contained in the Spotless
Maven plugin, but that was simple alphabetical sorting which is handled
using JavaParser in the Java TypeSpec generator. The Java TypeSpec
generator uses a mix of JavaParser and custom logic to prevent unwanted
JavaParser formatting.
- Update line ending usage to use newly added `Constants.NEW_LINE`,
which is always `\n` as git is configured for that to be the line
ending. This removes annoying fake file changes that are cleaned up by
git when running `git add --all`.

This change does bloat the size of the JAR shipped with Java TypeSpec
from about 8MB to 20MB which is acceptable as the size difference
would've been downloaded by Spotless when running anyways.