hsjs: fix handling of optional query/header params (#9176)
These use a different codepath in HSJS than JSON properties, and that
codepath was mishandling optional parameters by unconditionally passing
them into `globalThis.Number`/`globalThis.Boolean`, etc. This patch adds
an explicit check for `undefined` and transposes that to `undefined`
instead of calling the parser unconditionally, which will return the
default value. For parsers that don't accept `undefined` as an input,
this would have triggered a type error or runtime error.

Closes #9156